VR: move RelationAnalyzer contents to cpp
